// src/app/models/pupilDTO/pupilDTO.ts
export class PupilDTO {
  id: number;
  userId: number;
  name: string
  lastname: string;
  patronymic: string;
  dateOfBirthday: string;
  email: string;
  personalCheck: string;
  nameMom: string;
  lastnameMom: string;
  patronymicMom: string;
  nameDad: string;
  lastnameDad: string;
  patronymicDad: string;
  className: string;
}